fibonacci-omega/requirements.txt
# Core dependencies
numpy>=1.21.0
click>=8.0.0
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=2.0.0
pyyaml>=6.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-timeout>=2.1.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.990
